Install environment deps in root
This will install everything in the root conda environment.
I have also removed nb_conda and nb_conda_kernels so that
there is just one environment available and students don't
have to be exposed to the weirdness of having to select the
"root" or "default" environment, when these are in fact the
same.

@lwasser if you need other functionality from nb_conda, it
can be added again to the environment.yaml file!